requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
behave==1.2.6
ccxt==4.3.57
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.8
cssselect==1.2.0
exceptiongroup==1.2.1
execnet==2.1.1
fasteners==0.19
filelock==3.15.4
frozenlist==1.4.1
h11==0.14.0
idna==3.7
iniconfig==2.0.0
loguru==0.7.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.1
parameterized==0.9.0
parse==1.20.2
parse-type==0.6.2
pdbp==1.5.1
platformdirs==4.2.2
pluggy==1.5.0
py==1.11.0
pycares==4.4.0
pycparser==2.22
Pygments==2.18.0
pynose==1.5.1
pyotp==2.9.0
PySocks==1.7.1
pytest==8.2.1
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
sbvirtualdisplay==1.3.0
selenium==4.22.0
seleniumbase==4.28.3
setuptools==70.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tabcompleter==1.3.0
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wheel==0.43.0
wsproto==1.2.0
yarl==1.9.4
